Update README.md
Write a C program to find the second largest element in an array of integers. The program should prompt the user to enter the size of the array and the elements of the array, and then display the second largest element. If no second largest element exists (i.e., all elements are the same), the program should display an appropriate message